Split test event files to their respective folders.
